requirements.txt
pyserial>=3.5
tinyprog>=1.0.21